# openapi.yaml
openapi: 3.1.0
info:
  title: ContextGrip AI Chat
  description: >
    Self-hosted NL->SQL chat service. Ask questions about a PostgreSQL database
    in plain language; the service generates one read-only SQL statement,
    executes it inside a READ ONLY transaction with a statement timeout, and
    answers with the SQL always shown. Runs entirely on your own compute with
    your own Anthropic API key. This file is the source of truth for the
    server, the embedded UI, and the clients under clients/.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/status:
    get:
      tags: [system]
      summary: Service status for authenticated callers.
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/ask:
    post:
      tags: [chat]
      summary: One-shot question. Blocks until the full answer is ready and returns JSON.
      description: >
        Runs the same loop as the streaming endpoint (generate SQL -> verify
        read-only -> execute in READ ONLY tx -> explain) but returns a single
        JSON document. Creates a new conversation unless conversationId is
        given. Failed query execution is NOT an HTTP error: the response
        carries resultError and an answer explaining the failure.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AskRequest" }
      responses:
        "200":
          description: Answer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AskResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /api/v1/messages:
    post:
      tags: [chat]
      summary: Ask a question and stream the answer over SSE.
      description: >
        Server-Sent Events stream. Event sequence: meta -> sql -> result ->
        delta* -> done, or error terminally at any point after headers are
        sent. Every event's data field is a single JSON object. Creates a new
        conversation unless conversationId is given (the meta event carries
        the conversation id either way). Pre-stream failures (validation,
        auth, unknown conversation) are plain JSON errors with 4xx/5xx status.

        Event payloads:
          meta:   {"conversationId": string, "userMessageId": string}
          sql:    {"sql": string}
          result: ResultSummary, or {"error": string, "executionTimeMs": number} when execution failed
          delta:  {"text": string}
          done:   {"conversationId": string, "assistantMessageId": string, "pendingApprovalId"?: string}
          error:  {"message": string}

        Agent mode (AskRequest.mode = "agent", requires the "agent" feature):
        the model may take multiple tool steps before answering. Additional
        events, interleaved before the delta stream:
          step:              Step (one per completed tool step, in order)
          approval_required: Approval (a proposed write; the turn ends after
                             this — done carries pendingApprovalId — and the
                             write executes only via POST /api/v1/approvals/{id})
        In agent mode the sql/result pair may be absent (steps carry the
        queries instead).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AskRequest" }
      responses:
        "200":
          description: SSE stream (Content-Type text/event-stream)
          content:
            text/event-stream:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /api/v1/conversations:
    get:
      tags: [conversations]
      summary: List conversations, most recently updated first.
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Conversation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/conversations/{id}:
    get:
      tags: [conversations]
      summary: Fetch one conversation with its messages in order.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversationDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [conversations]
      summary: Delete a conversation and its messages.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties: { deleted: { type: boolean, const: true } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/approvals:
    get:
      tags: [approvals]
      summary: List pending write approvals (chat and board sources).
      responses:
        "200":
          description: Pending approvals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Approval" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/approvals/{id}:
    post:
      tags: [approvals]
      summary: Approve or reject a proposed write.
      description: >
        Approving executes the exact proposed SQL against the write
        connection (AI_CHAT_WRITE_DATABASE_URL) in a single transaction and
        returns the execution outcome; if no write connection is configured,
        approval fails with 409 WRITES_DISABLED. The decision is appended to
        the source conversation, and a board task waiting on this approval
        resumes automatically. Decisions are idempotent-once: deciding an
        already-decided approval returns 409 ALREADY_DECIDED.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approve, reject] }
      responses:
        "200":
          description: Decision recorded (and executed, when approved)
          content:
            application/json:
              schema:
                type: object
                required: [approval]
                properties:
                  approval: { $ref: "#/components/schemas/Approval" }
                  result: { $ref: "#/components/schemas/ResultSummary" }
                  error: { type: string, description: Execution error when the approved write failed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already decided, or writes not configured (code ALREADY_DECIDED | WRITES_DISABLED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: List board tasks, most recently updated first.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [queued, running, needs_approval, done, failed, canceled] }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [tasks]
      summary: File a task for the agent (requires the "board" feature).
      description: >
        Tasks run in the background through the same agent loop as agent-mode
        chat, one at a time, oldest first. Proposed writes pause the task in
        needs_approval; the approval decision resumes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, prompt]
              properties:
                title: { type: string, maxLength: 200 }
                prompt: { type: string, maxLength: 4000 }
      responses:
        "201":
          description: Filed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /api/v1/tasks/{id}:
    get:
      tags: [tasks]
      summary: Task detail — status, transcript steps, answer, pending approval.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [tasks]
      summary: Delete a finished task (done/failed/canceled only; 409 otherwise).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties: { deleted: { type: boolean, const: true } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: Task is still queued/running/needs_approval (code TASK_ACTIVE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/tasks/{id}/cancel:
    post:
      tags: [tasks]
      summary: Cancel a queued, running, or approval-blocked task.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Canceled task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already finished (code TASK_FINISHED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/messages/{id}/eval:
    post:
      tags: [training]
      summary: Rate an assistant answer (good/bad); writes a training record.
      description: >
        Explicit evals bypass the capture toggle — the caller deliberately
        asked to keep this record. Upserts by message id: rating the same
        answer again updates the verdict. Only assistant messages that carry
        SQL can be rated.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verdict]
              properties:
                verdict: { type: string, enum: [good, bad] }
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                properties: { recorded: { type: boolean, const: true } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/training/capture:
    get:
      tags: [training]
      summary: Read the automatic-capture setting.
      responses:
        "200":
          description: Setting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingCapture" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [training]
      summary: Enable/disable automatic capture (admin — primary APP_ACCESS_TOKEN only).
      description: >
        When enabled (the default), every completed chat exchange is recorded
        as a training record (question as intent, generated SQL, bounded
        result summary or execution error). Explicit evals are recorded
        regardless of this setting.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TrainingCapture" }
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingCapture" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminRequired" }
  /api/v1/training/stats:
    get:
      tags: [training]
      summary: Training-record counts and capture range.
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrainingStats" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/training/export:
    get:
      tags: [training]
      summary: Dump training records as JSONL (newline-delimited JSON).
      description: >
        Streams one TrainingExportLine JSON object per line
        (application/x-ndjson). The line format is compatible with
        ContextGrip's training-data export, so dumps from both sources merge
        downstream without transformation. The stream stops at a 64 MiB byte
        budget; compare line count with /api/v1/training/stats to detect
        truncation. Query params: includeRows (default true) includes bounded
        result row samples; evaluatedOnly (default false) restricts to
        records with an eval verdict.
      parameters:
        - { name: includeRows, in: query, schema: { type: boolean, default: true } }
        - { name: evaluatedOnly, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: JSONL stream
          content:
            application/x-ndjson:
              schema: { $ref: "#/components/schemas/TrainingExportLine" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/training/records:
    delete:
      tags: [training]
      summary: Delete ALL training records (admin — primary APP_ACCESS_TOKEN only).
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties: { deleted: { type: integer, description: Number of records removed } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminRequired" }
  /api/v1/tokens:
    get:
      tags: [tokens]
      summary: List named API tokens (admin — primary APP_ACCESS_TOKEN only).
      description: Raw token values are never returned; only label, fingerprint, and usage metadata.
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TokenInfo" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminRequired" }
    post:
      tags: [tokens]
      summary: Mint a named API token (admin — primary APP_ACCESS_TOKEN only).
      description: >
        The raw token value is returned once in this response and stored only
        as a SHA-256 hash. Named tokens can call every endpoint except token
        management.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string, maxLength: 120 }
      responses:
        "201":
          description: Created token (raw value shown once)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedToken" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminRequired" }
  /api/v1/tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke a named API token (admin — primary APP_ACCESS_TOKEN only).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties: { deleted: { type: boolean, const: true } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminRequired" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Either the primary APP_ACCESS_TOKEN from the server environment, or a
        named token minted via POST /api/v1/tokens. Token management endpoints
        accept only the primary token.
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token (code UNAUTHORIZED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AdminRequired:
      description: Named tokens cannot manage tokens (code ADMIN_REQUIRED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown resource (code NOT_FOUND)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotConfigured:
      description: Model key or database not configured (code NOT_CONFIGURED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable message }
        code:
          type: string
          description: >
            Stable machine slug. Known values: UNAUTHORIZED, ADMIN_REQUIRED,
            VALIDATION, NOT_FOUND, NOT_CONFIGURED, CONVERSATION_FULL,
            STORE_ERROR, MODEL_ERROR, STREAM_UNSUPPORTED, FEATURE_DISABLED
            (403 — surface not in AI_CHAT_FEATURES), WRITES_DISABLED,
            ALREADY_DECIDED, TASK_ACTIVE, TASK_FINISHED.
...
    Status:
      type: object
      required: [version, model, engine, ready, features, writesEnabled]
      properties:
        version: { type: string }
        model: { type: string, description: Anthropic model id in use }
        engine: { type: string, const: postgresql }
        ready: { type: boolean }
        features:
          type: array
          description: Enabled surfaces from AI_CHAT_FEATURES (chat, agent, board).
          items: { type: string, enum: [chat, agent, board] }
        writesEnabled:
          type: boolean
          description: True when AI_CHAT_WRITE_DATABASE_URL is configured — approvals can execute.
    AskRequest:
      type: object
      required: [question]
      properties:
        question: { type: string, maxLength: 4000 }
        conversationId:
          type: string
          description: Continue an existing conversation; omit to start a new one.
        mode:
          type: string
          enum: [chat, agent]
          default: chat
          description: >
            agent lets the model take multiple tool steps (read-only queries
            run automatically; writes become approvals). Requires the "agent"
            feature; a conversation keeps the mode of its first message.
    AskResponse:
      type: object
      required: [conversationId, userMessageId, assistantMessageId, answer]
      properties:
        conversationId: { type: string }
        userMessageId: { type: string }
        assistantMessageId: { type: string }
        sql: { type: string, description: "The generated read-only SQL (chat mode; agent mode may carry steps instead)" }
        result: { $ref: "#/components/schemas/ResultSummary" }
        resultError:
          type: string
          description: Set instead of result when query execution failed.
        answer: { type: string, description: Natural-language explanation of the outcome }
        steps:
          type: array
          description: Agent-mode tool steps, in execution order.
          items: { $ref: "#/components/schemas/Step" }
        pendingApprovalId:
          type: string
          description: Set when the turn ended awaiting a write approval.
    Step:
      type: object
      required: [index, kind, summary]
      properties:
        index: { type: integer }
        kind: { type: string, enum: [query, schema, note] }
        summary: { type: string, description: One-line description of the step }
        sql: { type: string, description: The read-only SQL this step executed (kind query) }
        result: { $ref: "#/components/schemas/ResultSummary" }
        error: { type: string, description: Step execution error }
    Approval:
      type: object
      required: [id, sql, status, source, createdAt]
      properties:
        id: { type: string }
        sql: { type: string, description: The exact statement that will run if approved }
        rationale: { type: string, description: The model's stated reason for the write }
        status: { type: string, enum: [pending, approved, rejected] }
        source:
          type: object
          description: Where the proposal came from (exactly one of the ids is set).
          properties:
            conversationId: { type: string }
            messageId: { type: string }
            taskId: { type: string }
        createdAt: { type: string, format: date-time }
        decidedAt: { type: string, format: date-time }
    Task:
      type: object
      required: [id, title, prompt, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        prompt: { type: string }
        status: { type: string, enum: [queued, running, needs_approval, done, failed, canceled] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        answer: { type: string, description: Final answer when done }
        error: { type: string, description: Failure reason when failed }
    TaskDetail:
      type: object
      required: [task, steps]
      properties:
        task: { $ref: "#/components/schemas/Task" }
        steps:
          type: array
          items: { $ref: "#/components/schemas/Step" }
        pendingApproval: { $ref: "#/components/schemas/Approval" }
    ResultSummary:
      type: object
      required: [rowCount, truncated, executionTimeMs]
      properties:
        columns:
          type: array
          items: { type: string }
        rowSample:
          type: array
          description: Up to 20 rows; string cells bounded to 256 chars.
          items:
            type: array
            items: {}
        rowCount: { type: integer, description: Rows returned (capped at 100; see truncated) }
        truncated: { type: boolean }
        executionTimeMs: { type: integer }
    Conversation:
      type: object
      required: [id, title, mode, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string, description: Derived from the first question, max 80 chars }
        mode:
          type: string
          enum: [chat, agent]
          description: Fixed by the conversation's first message.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ConversationDetail:
      type: object
      required: [conversation, messages]
      properties:
        conversation: { $ref: "#/components/schemas/Conversation" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/Message" }
...
    TrainingCapture:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
    TrainingStats:
      type: object
      required: [records, evaluated]
      properties:
        records: { type: integer }
        evaluated: { type: integer, description: Records carrying an eval verdict }
        firstCapturedAt: { type: string, format: date-time }
        lastCapturedAt: { type: string, format: date-time }
    TrainingExportLine:
      type: object
      description: One JSONL line. Field layout matches ContextGrip's training export.
      required: [id, capturedAt, connection, context, query, response]
      properties:
        id: { type: string }
        capturedAt: { type: string, format: date-time }
        connection:
          type: object
          required: [id, name, engine]
          properties:
            id: { type: string, description: Stable non-secret hash of host:port/dbname }
            name: { type: string, description: Database name from DATABASE_URL }
            engine: { type: string, const: postgresql }
        context:
          type: object
          properties:
            session: { type: string, description: "\"chat\" (SSE) or \"ask\" (one-shot)" }
            sourceMessageId: { type: string, description: Assistant message id, for dedupe }
        query:
          type: object
          required: [sql]
          properties:
            sql: { type: string }
            intent: { type: string, description: The natural-language question }
        response:
          type: object
          required: [rowCount, truncated, executionTimeMs]
          properties:
            columns: { type: array, items: { type: string } }
            rowCount: { type: integer }
            truncated: { type: boolean }
            executionTimeMs: { type: integer }
            error: { type: string }
            rowSample:
              type: array
              items: { type: array, items: {} }
        eval:
          type: object
          required: [verdict]
          properties:
            verdict: { type: string, enum: [good, bad] }
    TokenInfo:
      type: object
      required: [id, label, fingerprint, createdAt]
      properties:
        id: { type: string }
        label: { type: string }
        fingerprint: { type: string, description: First 8 hex chars of SHA-256(token) }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time }
    CreatedToken:
      allOf:
        - $ref: "#/components/schemas/TokenInfo"
        - type: object
          required: [token]
          properties:
            token: { type: string, description: Raw token value — shown only in this response }